structure eval util + attempt evaluate
